+++ src/leader.js
@@ -2,21 +2,20 @@ class AvailableLeaders {
     constructor() {
         this.allLeaders = {};
         this.availableLeaders = {};
-        this.addLeader("Gleb");
-        this.addLeader("Sviatoslav");
-        this.addLeader("Theofana");
-        this.addLeader("Agatha");
-        this.addLeader("Sudislav");
-        this.addLeader("Boris");
-        this.addLeader("Sviatopolk");
-        this.addLeader("Maria");
-        this.addLeader("Yaroslav");
-        this.addLeader("Mstislav");
-        this.addLeader("Predslava");
-    }
+        this.addLeader("Boris", "When you attack an opponent in Boris's region, reveal one less scheme card and steal one coin from that opponent.");
+        this.addLeader("Sviatopolk", "Rebels in Sviatopolk's region count as your troops for purposes of rule. If you defeat a rebel in that region, replace it with one of your troops.");
+        this.addLeader("Yaroslav", "", "In Yaroslav's region, you win ties for rule and your opponents lose the abilities of their structures.");
+        this.addLeader("Mstislav", "In Mstislav's region, it only costs you one tax point to tax or one build point to build, regardless of who rules that region.");
+        this.addLeader("Gleb", "", "Once per round, when you attack in Gleb's region with no casualty, gain two movement points to use with any of your troops located in his region (including Gleb).");
+        this.addLeader("Theofana", "", "Once per round, when you tax in Theofana's region, you may move her to an adjacent region or gain one coin from the supply. If you use this ability, you cannot tax again this turn.");
+        this.addLeader("Maria", "", "Once per round when you muster troops, you may place them in a single region adjacent to Maria's region that you do not occupy.");
+        this.addLeader("Predslava", "", "Once per round, on your turn, you may move an opponent's troop to an adjacent region for free. The opponent gains one coin from the supply.");
+        this.addLeader("Sviatoslav", "", "Before you spend any build points on your turn, you may move Sviatoslav to an adjacent region with a rebel or spend one coin to move Sviatoslav to any adjacent region.");
+        this.addLeader("Agatha", "", "When you move Agatha, you may move up to two of your troops along with her for free.");
+        this.addLeader("Sudislav", "", "You may spend attack points as muster points instead, to muster troops in Sudislav's region.");    }
 
-    addLeader(name) {
-        var leader = new Leader(name);
+    addLeader(name, description=null) {
+        var leader = new Leader(name, description);
         this.availableLeaders[name] = leader;
         this.allLeaders[name] = leader;
     }
@@ -43,8 +42,9 @@ class AvailableLeaders {
 
 
 class Leader {
-    constructor(name) {
+    constructor(name, description=null) {
         this.name = name;
+        this.description = description;
     }
 }